requirements.txt
streamlit
langchain
faiss-cpu
sentence-transformers
pypdf
python-docx
python-pptx
